+++ zcash_client_backend/src/data_api/chain.rs
@@ -151,8 +151,7 @@ use zcash_primitives::consensus::{self, BlockHeight};
 use crate::{
     data_api::{NullifierQuery, WalletWrite},
     proto::compact_formats::CompactBlock,
-    scan::BatchRunner,
-    scanning::{add_block_to_runner, scan_block_with_runner, ScanningKeys},
+    scanning::{scan_block_with_runners, BatchRunners, ScanningKeys},
 };
 
 pub mod error;
@@ -216,22 +215,26 @@ pub trait BlockSource {
 /// blocks.
 #[derive(Clone, Debug)]
 pub struct ScanSummary {
-    scanned_range: Range<BlockHeight>,
-    spent_sapling_note_count: usize,
-    received_sapling_note_count: usize,
+    pub(crate) scanned_range: Range<BlockHeight>,
+    pub(crate) spent_sapling_note_count: usize,
+    pub(crate) received_sapling_note_count: usize,
+    #[cfg(feature = "orchard")]
+    pub(crate) spent_orchard_note_count: usize,
+    #[cfg(feature = "orchard")]
+    pub(crate) received_orchard_note_count: usize,
 }
 
 impl ScanSummary {
-    /// Constructs a new [`ScanSummary`] from its constituent parts.
-    pub fn from_parts(
-        scanned_range: Range<BlockHeight>,
-        spent_sapling_note_count: usize,
-        received_sapling_note_count: usize,
-    ) -> Self {
+    /// Constructs a new [`ScanSummary`] for the provided block range.
+    pub(crate) fn for_range(scanned_range: Range<BlockHeight>) -> Self {
         Self {
             scanned_range,
-            spent_sapling_note_count,
-            received_sapling_note_count,
+            spent_sapling_note_count: 0,
+            received_sapling_note_count: 0,
+            #[cfg(feature = "orchard")]
+            spent_orchard_note_count: 0,
+            #[cfg(feature = "orchard")]
+            received_orchard_note_count: 0,
         }
     }
 
@@ -248,12 +251,30 @@ impl ScanSummary {
         self.spent_sapling_note_count
     }
 
-    /// Returns the number of notes belonging to the wallet that were received in blocks in the
-    /// scanned range. Note that depending upon the scanning order, it is possible that some of the
-    /// received notes counted here may already have been spent in later blocks closer to the chain
-    /// tip.
+    /// Returns the number of Sapling notes belonging to the wallet that were received in blocks in
+    /// the scanned range. Note that depending upon the scanning order, it is possible that some of
+    /// the received notes counted here may already have been spent in later blocks closer to the
+    /// chain tip.
     pub fn received_sapling_note_count(&self) -> usize {
         self.received_sapling_note_count
+    }
+
+    /// Returns the number of our previously-detected Orchard notes that were spent in transactions
+    /// in blocks in the scanned range. If we have not yet detected a particular note as ours, for
+    /// example because we are scanning the chain in reverse height order, we will not detect it
+    /// being spent at this time.
+    #[cfg(feature = "orchard")]
+    pub fn spent_orchard_note_count(&self) -> usize {
+        self.spent_orchard_note_count
+    }
+
+    /// Returns the number of Orchard notes belonging to the wallet that were received in blocks in
+    /// the scanned range. Note that depending upon the scanning order, it is possible that some of
+    /// the received notes counted here may already have been spent in later blocks closer to the
+    /// chain tip.
+    #[cfg(feature = "orchard")]
+    pub fn received_orchard_note_count(&self) -> usize {
+        self.received_orchard_note_count
     }
 }
 
@@ -279,11 +300,10 @@ where
     <DbT as WalletRead>::AccountId: ConditionallySelectable + Default + Send + 'static,
 {
     // Fetch the UnifiedFullViewingKeys we are tracking
-    let mut scanning_keys = ScanningKeys::from_account_ufvks(
-        data_db
-            .get_unified_full_viewing_keys()
-            .map_err(Error::Wallet)?,
-    );
+    let account_ufvks = data_db
+        .get_unified_full_viewing_keys()
+        .map_err(Error::Wallet)?;
+    let mut scanning_keys = ScanningKeys::from_account_ufvks(account_ufvks);
 
     // Get the nullifiers for the unspent notes we are tracking
     scanning_keys.extend_sapling_nullifiers(
@@ -291,26 +311,26 @@ where
             .get_sapling_nullifiers(NullifierQuery::Unspent)
             .map_err(Error::Wallet)?,
     );
-
-    let mut sapling_runner = BatchRunner::<_, _, _, _, ()>::new(
-        100,
-        scanning_keys
-            .sapling_keys()
-            .iter()
-            .map(|(id, key)| (*id, key.prepare())),
+    #[cfg(feature = "orchard")]
+    scanning_keys.extend_orchard_nullifiers(
+        data_db
+            .get_orchard_nullifiers(NullifierQuery::Unspent)
+            .map_err(Error::Wallet)?,
     );
+
+    let mut runners = BatchRunners::<_, (), ()>::for_keys(100, &scanning_keys);
 
     block_source.with_blocks::<_, DbT::Error>(
         Some(from_height),
         Some(limit),
         |block: CompactBlock| {
-            add_block_to_runner(params, block, &mut sapling_runner);
+            runners.add_block(params, block);
 
             Ok(())
         },
     )?;
 
-    sapling_runner.flush();
+    runners.flush();
 
     let mut prior_block_metadata = if from_height > BlockHeight::from(0) {
         data_db
@@ -321,42 +341,37 @@ where
     };
 
     let mut scanned_blocks = vec![];
-    let mut scan_end_height = from_height;
-    let mut received_note_count = 0;
-    let mut spent_note_count = 0;
+    let mut scan_summary = ScanSummary::for_range(from_height..from_height);
     block_source.with_blocks::<_, DbT::Error>(
         Some(from_height),
         Some(limit),
         |block: CompactBlock| {
-            scan_end_height = block.height() + 1;
-            let scanned_block = scan_block_with_runner(
+            scan_summary.scanned_range.end = block.height() + 1;
+            let scanned_block = scan_block_with_runners::<_, _, _, (), ()>(
                 params,
                 block,
                 &scanning_keys,
                 prior_block_metadata.as_ref(),
-                Some(&mut sapling_runner),
+                Some(&mut runners),
             )
             .map_err(Error::Scan)?;
 
-            let (s, r) = scanned_block
-                .transactions
-                .iter()
-                .fold((0, 0), |(s, r), wtx| {
-                    (
-                        s + wtx.sapling_spends().len(),
-                        r + wtx.sapling_outputs().len(),
-                    )
-                });
-            spent_note_count += s;
-            received_note_count += r;
+            for wtx in &scanned_block.transactions {
+                scan_summary.spent_sapling_note_count += wtx.sapling_spends().len();
+                scan_summary.received_sapling_note_count += wtx.sapling_outputs().len();
+                #[cfg(feature = "orchard")]
+                {
+                    scan_summary.spent_orchard_note_count += wtx.orchard_spends().len();
+                    scan_summary.received_orchard_note_count += wtx.orchard_outputs().len();
+                }
+            }
 
-            let spent_nf: Vec<&sapling::Nullifier> = scanned_block
+            let sapling_spent_nf: Vec<&sapling::Nullifier> = scanned_block
                 .transactions
                 .iter()
                 .flat_map(|tx| tx.sapling_spends().iter().map(|spend| spend.nf()))
                 .collect();
-
-            scanning_keys.retain_sapling_nullifiers(|(_, nf)| !spent_nf.contains(&nf));
+            scanning_keys.retain_sapling_nullifiers(|(_, nf)| !sapling_spent_nf.contains(&nf));
             scanning_keys.extend_sapling_nullifiers(scanned_block.transactions.iter().flat_map(
                 |tx| {
                     tx.sapling_outputs()
@@ -364,6 +379,24 @@ where
                         .flat_map(|out| out.nf().into_iter().map(|nf| (*out.account_id(), *nf)))
                 },
             ));
+
+            #[cfg(feature = "orchard")]
+            {
+                let orchard_spent_nf: Vec<&orchard::note::Nullifier> = scanned_block
+                    .transactions
+                    .iter()
+                    .flat_map(|tx| tx.orchard_spends().iter().map(|spend| spend.nf()))
+                    .collect();
+
+                scanning_keys.retain_orchard_nullifiers(|(_, nf)| !orchard_spent_nf.contains(&nf));
+                scanning_keys.extend_orchard_nullifiers(
+                    scanned_block.transactions.iter().flat_map(|tx| {
+                        tx.orchard_outputs()
+                            .iter()
+                            .flat_map(|out| out.nf().into_iter().map(|nf| (*out.account_id(), *nf)))
+                    }),
+                );
+            }
 
             prior_block_metadata = Some(scanned_block.to_block_metadata());
             scanned_blocks.push(scanned_block);
@@ -373,11 +406,7 @@ where
     )?;
 
     data_db.put_blocks(scanned_blocks).map_err(Error::Wallet)?;
-    Ok(ScanSummary::from_parts(
-        from_height..scan_end_height,
-        spent_note_count,
-        received_note_count,
-    ))
+    Ok(scan_summary)
 }
 
 #[cfg(feature = "test-dependencies")]